requirements.txt
fastapi~=0.112.2
uvicorn~=0.30.6
tqdm~=4.66.5
python-dotenv~=1.0.1
protobuf~=5.28.0
pymongo~=4.8.0